Consolidate workflow run ownership
Move manual workflow-run firing and optimistic list cache updates into @shipfox/client-workflows so the project workflows page consumes the same cache owner as the run rail.

Fixes ENG-514
