+++ .changeset/central-runs-fire.md
@@ -1,0 +1,6 @@
+---
+"@shipfox/client-projects": patch
+"@shipfox/client-workflows": patch
+---
+
+Moves manual workflow-run firing and optimistic run-list cache updates into `@shipfox/client-workflows` so project workflow pages consume the run cache owner directly.
